Child plan sections extraction
